Make it easier to find Emacs User Guide in Docs (#264)
If you navigate to the `/editors/emacs` page it's a bit hard to find the User's Guide, since the sidebar flushes content to the very bottom of the browser frame, it's practically invisible on a tall, vertically oriented monitor. This adds a sentence pointing users to the User's Guide from that page.